+++ .gitignore
@@ -1,2 +1,4 @@
+.cache
 *.code-workspace
+dist
 node_modules
