+++ index.html
@@ -7,7 +7,7 @@
 		<link href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
 		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 		<link href="style.css" rel="stylesheet"> 
-		<script src="app.js" defer></script>
+		<script src="scripts/app.js" defer></script>
 	</head>
 	<body>
 		<div class="jumbotron text-center">
