--- test/app.spec.js
@@ -1,7 +1,0 @@
-const assert = require('assert');
-
-describe('app.js', () => {
-    it('1 plus 1 is 2', () => {
-        assert.equal(1, 1);
-    });
-});
