+++ test/cipher.spec.js
@@ -1,0 +1,8 @@
+const assert = require('assert');
+import Cipher from '../scripts/cipher';
+
+describe('cipher.js', () => {
+    it('1 plus 1 is 2', () => {
+        console.log(Cipher);
+    });
+});
